moved the tests for button
